+++ Controllers/RecipeController.cs
@@ -1,0 +1,10 @@
+using Microsoft.AspNetCore.Mvc;
+
+namespace Reciplease.Controllers;
+
+[ApiController]
+[Route("[controller]")]
+public class RecipeController : ControllerBase
+{
+
+}
